remove file

+++ src/midi/midi.h
@@ -8,9 +8,9 @@
 typedef struct {
     unsigned char MTHD[4]; // 4 bytes for header flag
     unsigned char SPEC_ADDR[4]; // 4 bytes to inform about the addressing specs
-    unsigned int SMF ; // 2 bytes to indicate the data ranges: 0 only one data range, 1 several tracks played simultaneously, 2 more tracks but to be played one after the other
-    unsigned int MTRK ; // 2 bytes for the number of data ranges
-    unsigned int NOIRE; // 2 bytes: number divisions of the black
+    u_int16_t SMF ; // 2 bytes to indicate the data ranges: 0 only one data range, 1 several tracks played simultaneously, 2 more tracks but to be played one after the other
+    u_int16_t MTRK ; // 2 bytes for the number of data ranges
+    u_int16_t NOIRE; // 2 bytes: number divisions of the black
 
 } Header;
